samples update
