+++ keyvault/client/keys/src/samples/java/com/azure/security/keyvault/keys/BackupAndRestoreOperationsAsync.java
@@ -1,11 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package keys;
+package com.azure.security.keyvault.keys;
 
 import com.azure.identity.credential.DefaultAzureCredential;
-import com.azure.security.keyvault.keys.KeyClient;
-import com.azure.security.keyvault.keys.models.Key;
 import com.azure.security.keyvault.keys.models.RsaKeyCreateOptions;
 
 import java.io.File;
@@ -16,11 +14,11 @@ import java.nio.file.Files;
 import java.time.OffsetDateTime;
 
 /**
- * Sample demonstrates how to backup and restore keys in the key vault.
+ * Sample demonstrates how to asynchronously backup and restore keys in the key vault.
  */
-public class BackupAndRestoreOperations {
+public class BackupAndRestoreOperationsAsync {
     /**
-     * Authenticates with the key vault and shows how to backup and restore keys in the key vault.
+     * Authenticates with the key vault and shows how to asynchronously backup and restore keys in the key vault.
      *
      * @param args Unused. Arguments to the program.
      * @throws IllegalArgumentException when invalid key vault endpoint is passed.
@@ -29,41 +27,55 @@ public class BackupAndRestoreOperations {
      */
     public static void main(String[] args) throws IOException, InterruptedException, IllegalArgumentException {
 
-        // Instantiate a client that will be used to call the service. Notice that the client is using default Azure
+        // Instantiate async key client that will be used to call the service. Notice that the client is using default Azure
         // credentials. To make default credentials work, ensure that environment variables 'AZURE_CLIENT_ID',
         // 'AZURE_CLIENT_KEY' and 'AZURE_TENANT_ID' are set with the service principal credentials.
-        KeyClient keyClient = KeyClient.builder()
-                .endpoint("https://{YOUR_VAULT_NAME}.vault.azure.net")
-                .credential(new DefaultAzureCredential())
-                .build();
+        KeyAsyncClient keyAsyncClient = KeyAsyncClient.builder()
+            .endpoint("https://{YOUR_VAULT_NAME}.vault.azure.net")
+            .credential(new DefaultAzureCredential())
+            .build();
 
-        // Let's create a Rsa key valid for 1 year. if the key
+        // Let's create Cloud Rsa key valid for 1 year. if the key
         // already exists in the key vault, then a new version of the key is created.
-        keyClient.createRsaKey(new RsaKeyCreateOptions("CloudRsaKey")
+        keyAsyncClient.createRsaKey(new RsaKeyCreateOptions("CloudRsaKey")
                 .expires(OffsetDateTime.now().plusYears(1))
-                .keySize(2048));
+                .keySize(2048))
+                .subscribe(keyResponse ->
+                        System.out.printf("Key is created with name %s and type %s \n", keyResponse.value().name(), keyResponse.value().keyMaterial().kty()));
+
+        Thread.sleep(2000);
 
         // Backups are good to have, if in case keys get accidentally deleted by you.
         // For long term storage, it is ideal to write the backup to a file.
         String backupFilePath = "YOUR_BACKUP_FILE_PATH";
-        byte[] keyBackup = keyClient.backupKey("CloudRsaKey").value();
-        writeBackupToFile(keyBackup, backupFilePath);
+        keyAsyncClient.backupKey("CloudRsaKey").subscribe(backupResponse -> {
+            byte[] backupBytes = backupResponse.value();
+            writeBackupToFile(backupBytes, backupFilePath);
+        });
+
+        Thread.sleep(7000);
 
         // The Cloud Rsa key is no longer in use, so you delete it.
-        keyClient.deleteKey("CloudRsaKey");
+        keyAsyncClient.deleteKey("CloudRsaKey").subscribe(deletedKeyResponse ->
+                System.out.printf("Deleted Key's Recovery Id %s \n", deletedKeyResponse.value().recoveryId()));
 
-        //To ensure key is deleted on server side.
+        //To ensure file is deleted on server side.
         Thread.sleep(30000);
 
         // If the vault is soft-delete enabled, then you need to purge the key as well for permanent deletion.
-        keyClient.purgeDeletedKey("CloudRsaKey");
+        keyAsyncClient.purgeDeletedKey("CloudRsaKey").subscribe(purgeResponse ->
+            System.out.printf("Purge Status response %d \n", purgeResponse.statusCode()));
 
-        //To ensure key is purged on server side.
+        //To ensure file is purged on server side.
         Thread.sleep(15000);
 
         // After sometime, the key is required again. We can use the backup value to restore it in the key vault.
         byte[] backupFromFile = Files.readAllBytes(new File(backupFilePath).toPath());
-        Key restoredKey = keyClient.restoreKey(backupFromFile).value();
+        keyAsyncClient.restoreKey(backupFromFile).subscribe(keyResponse ->
+            System.out.printf("Restored Key with name %s \n", keyResponse.value().name()));
+
+        //To ensure key is restored on server side.
+        Thread.sleep(15000);
     }
 
     private static void writeBackupToFile(byte[] bytes, String filePath) {
