+++ keyvault/client/secrets/src/samples/java/com/azure/security/keyvault/secrets/HelloWorldAsync.java
@@ -1,10 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package secrets;
+package com.azure.security.keyvault.secrets;
 
 import com.azure.identity.credential.DefaultAzureCredential;
-import com.azure.security.keyvault.secrets.SecretAsyncClient;
 import com.azure.security.keyvault.secrets.models.Secret;
 import java.time.OffsetDateTime;
 
@@ -32,8 +31,8 @@ public class HelloWorldAsync {
         // Let's create a secret holding bank account credentials valid for 1 year. if the secret
         // already exists in the key vault, then a new version of the secret is created.
         secretAsyncClient.setSecret(new Secret("BankAccountPassword", "f4G34fMh8v")
-         .expires(OffsetDateTime.now().plusYears(1))).subscribe(secretResponse ->
-             System.out.printf("Secret is created with name %s and value %s \n", secretResponse.value().name(), secretResponse.value().value()));
+            .expires(OffsetDateTime.now().plusYears(1))).subscribe(secretResponse ->
+                System.out.printf("Secret is created with name %s and value %s \n", secretResponse.value().name(), secretResponse.value().value()));
 
         Thread.sleep(2000);
 
@@ -47,11 +46,11 @@ public class HelloWorldAsync {
         // The update method can be used to update the expiry attribute of the secret. It cannot be used to update
         // the value of the secret.
         secretAsyncClient.getSecret("BankAccountPassword").subscribe(secretResponse -> {
-          Secret secret = secretResponse.value();
-          //Update the expiry time of the secret.
-          secret.expires(secret.expires().plusYears(1));
-          secretAsyncClient.updateSecret(secret).subscribe(updatedSecretResponse ->
-              System.out.printf("Secret's updated expiry time %s \n", updatedSecretResponse.value().expires().toString()));
+            Secret secret = secretResponse.value();
+            //Update the expiry time of the secret.
+            secret.expires(secret.expires().plusYears(1));
+            secretAsyncClient.updateSecret(secret).subscribe(updatedSecretResponse ->
+                System.out.printf("Secret's updated expiry time %s \n", updatedSecretResponse.value().expires().toString()));
         });
 
         Thread.sleep(2000);
@@ -60,20 +59,20 @@ public class HelloWorldAsync {
         // To achieve this, we need to create a new version of the secret in the key vault. The update operation cannot
         // change the value of the secret.
         secretAsyncClient.setSecret("BankAccountPassword", "bhjd4DDgsa").subscribe(secretResponse ->
-          System.out.printf("Secret is created with name %s and value %s \n", secretResponse.value().name(), secretResponse.value().value()));
+            System.out.printf("Secret is created with name %s and value %s \n", secretResponse.value().name(), secretResponse.value().value()));
 
         Thread.sleep(2000);
 
         // The bank account was closed, need to delete its credentials from the key vault.
         secretAsyncClient.deleteSecret("BankAccountPassword").subscribe(deletedSecretResponse ->
-          System.out.printf("Deleted Secret's Recovery Id %s \n", deletedSecretResponse.value().recoveryId()));
+            System.out.printf("Deleted Secret's Recovery Id %s \n", deletedSecretResponse.value().recoveryId()));
 
         //To ensure secret is deleted on server side.
         Thread.sleep(30000);
 
         // If the key vault is soft-delete enabled, then for permanent deletion deleted secrets need to be purged.
         secretAsyncClient.purgeDeletedSecret("BankAccountPassword").subscribe(purgeResponse ->
-                System.out.printf("Bank account secret purge status response %d \n", purgeResponse.statusCode()));
+            System.out.printf("Bank account secret purge status response %d \n", purgeResponse.statusCode()));
 
         //To ensure secret is purged on server side.
         Thread.sleep(15000);
